--- src/README.md
@@ -1,25 +1,0 @@
-### Lazy Chrome
-
-Lazy Chrome is a mouse terminator
-A keyboard short-cut customization tool with many features of operate tab
-
-Super speed up your common operations with your custom shortcut keys
-duplicate, pin or paste URL and open in a new tab!
-*********
-[Lazy Chrome - Chrome 線上應用程式商店](https://chrome.google.com/webstore/detail/lazy-chrome/boijmdgjfcoancndoiclnglhjakeopch "Lazy Chrome - Chrome 線上應用程式商店")
-
-## Features
-click the app icon to config your short cuts
-the default actions:
-
-- pin: alt + 1
-- duplicate: alt + 3
-- paste and open: ctrl + shift + v
-- paste and search by Google: ctrl + shift + v
-- search with selected text
-- make a tab be independent
-- open notepad in Chrome
-- mute current tab
-- mute the other tabs
-- prevent window close
-- only keep the same domain tabs
