--- src/background.js
@@ -1,337 +1,0 @@
-import {Common} from "./common.js"
-// Copyright 2018 The Chromium Authors. All rights reserved.
-// Use of this source code is governed by a BSD-style license that can be
-// found in the LICENSE file.
-
-'use strict';
-Object.prototype.let = function (callback) {
-  return callback.apply(this, [this])
-}
-
-chrome.browserAction.onClicked.addListener(tab => {
-  console.log('browserAction', tab)
-  // chrome.tabs.create({
-  //   url: "chrome://extensions/shortcuts"
-  // })
-});
-
-var curTabID = 0;
-var windowsHistory = []
-
-chrome.tabs.onActivated.addListener(info => {
-  console.log('onActivated', info)
-  var target = {}
-  var targetWindow = windowsHistory.find(element => element.windowId == info.windowId)
-  if (targetWindow) {
-    targetWindow.lastId = targetWindow.tabId
-    targetWindow.tabId = info.tabId
-    target = targetWindow
-  } else {
-    windowsHistory.push(info)
-    target = info
-  }
-
-  chrome.runtime.sendMessage({
-    activatedObj: target
-  });
-})
-
-var windowsOnCreatedTimeount
-chrome.windows.onRemoved.addListener(windowId => {
-  console.log('windows.onRemoved', windowId)
-  clearTimeout(windowsOnCreatedTimeount)
-})
-
-chrome.windows.onCreated.addListener(window => {
-  console.log('windows.onCreated', window)
-  Common.isPreventClose(value => {
-    if (value && window.type == 'normal') {
-      windowsOnCreatedTimeount = setTimeout(() => {
-        try {
-          chrome.tabs.query({
-            windowId: window.id
-          }, tabs => {
-            tabs.forEach(tab => {
-              tab.url == Common.getPreventCloseTabUrl() && chrome.tabs.remove(tab.id)
-            })
-            chrome.tabs.create({
-              windowId: window.id,
-              url: Common.getPreventCloseTabUrl(),
-              pinned: true
-            }, _ => {
-              clearTimeout(windowsOnCreatedTimeount)
-            })
-          })
-        } catch (error) {
-          console.error('windows.onCreated', error)
-          clearTimeout(windowsOnCreatedTimeount)
-        }
-      }, 666)
-    }
-  })
-})
-
-chrome.runtime.onInstalled.addListener(function () {
-  console.log('onInstalled')
-  // chrome.storage.sync.set({
-  //   color: '#3aa757'
-  // }, function () {
-  //   console.log("The color is green.");
-  // });
-
-  chrome.declarativeContent.onPageChanged.removeRules(undefined, function () {
-    chrome.declarativeContent.onPageChanged.addRules([{
-      conditions: [new chrome.declarativeContent.PageStateMatcher({
-        pageUrl: {
-          hostEquals: 'developer.chrome.com'
-        },
-      })],
-      actions: [new chrome.declarativeContent.ShowPageAction()]
-    }]);
-  });
-});
-
-class Commands {
-  static duplicate() {
-    Common.getCurrentTab(tabs => {
-      var current = tabs[0]
-      chrome.tabs.duplicate(current.id);
-    });
-  }
-
-  static toShutUp() {
-    chrome.tabs.query({
-      currentWindow: true,
-      audible: true
-    }, tabs => {
-      Common.getCurrentTab(current => {
-        tabs.forEach(element => {
-          chrome.tabs.update(element.id, {
-            'muted': element.id == current[0].id ? current[0].mutedInfo.muted : true
-          });
-        })
-      })
-    });
-  }
-
-  static toggleMute() {
-    Common.getCurrentTab(tabs => {
-      var current = tabs[0]
-      console.log(current)
-      chrome.tabs.update(current.id, {
-        'muted': !current.mutedInfo.muted
-      });
-    });
-  }
-
-  static togglePin() {
-    Common.getCurrentTab(tabs => {
-      var current = tabs[0]
-      chrome.windows.create({
-        focused: true,
-        type: 'normal'
-      }, newWin => {
-        chrome.tabs.move(current.id, {
-          windowId: newWin.id,
-          index: -1
-        }, _ => {
-          Common.getCurrentTab(tabs => {
-            chrome.tabs.remove(tabs[0].id);
-          })
-        });
-      })
-    });
-  }
-  static openNotepad() {
-    chrome.tabs.create({
-      url: "data:text/html, <html contenteditable>"
-    })
-  }
-  static newQueryWithSelected() {
-    chrome.tabs.executeScript({
-      code: "window.getSelection().toString();"
-    }, selection => {
-      var clipboardContents = selection[0]
-      chrome.tabs.create({
-        url: "https://www.google.com/search?q=" + (clipboardContents || "")
-      })
-    });
-  }
-  static newQueryWithPasted() {
-    var clipboardContents = getPasted()
-    chrome.tabs.create({
-      url: "https://www.google.com/search?q=" + (clipboardContents || "")
-    })
-  }
-  static newTabWithUrl() {
-    var clipboardContents = getPasted()
-    var isUrl = clipboardContents.indexOf("://") > -1
-    chrome.tabs.create({
-      url: (isUrl ? clipboardContents : "https://www.google.com/search?q=" + (clipboardContents || "")) || "https://www.google.com/"
-    })
-  }
-  static keepSameDomain() {
-    Common.getCurrentTab(currentTab => {
-      var preventCloseTabUrl = Common.getPreventCloseTabUrl()
-      if (currentTab[0].url == preventCloseTabUrl) return
-      var currentDomain = (new URL(currentTab[0].url)).hostname
-      try {
-        Common.isPreventClose(value => {
-          chrome.tabs.query({
-            currentWindow: true,
-            pinned: false
-          }, tabs => {
-            [].let(it => {
-              tabs.forEach(tab => {
-                var domain = (new URL(tab.url)).hostname
-                if (domain != currentDomain &&
-                  ((value && tab.url != preventCloseTabUrl) ||
-                    !value)) {
-                  it.push(tab.id)
-                }
-              })
-              it.length > 0 && chrome.tabs.remove(it)
-            })
-          })
-        })
-      } catch (error) {
-        console.error(error)
-      }
-    })
-  }
-  static killOtherSameDomain() {
-    Common.getCurrentTab(currentTab => {
-      var preventCloseTabUrl = Common.getPreventCloseTabUrl()
-      if (currentTab[0].url == preventCloseTabUrl) return
-      var currentDomain = (new URL(currentTab[0].url)).hostname
-      try {
-        Common.isPreventClose(value => {
-          chrome.tabs.query({
-            currentWindow: true,
-            pinned: false
-          }, tabs => {
-            [].let(it => {
-              tabs.forEach(tab => {
-                var domain = (new URL(tab.url)).hostname
-                if (tab.id != currentTab[0].id &&
-                  !(domain != currentDomain &&
-                    ((value && tab.url != preventCloseTabUrl) ||
-                      !value))) {
-                  it.push(tab.id)
-                }
-              })
-              it.length > 0 && chrome.tabs.remove(it)
-            })
-          })
-        })
-      } catch (error) {
-        console.error(error)
-      }
-    })
-  }
-  static killSameDomain() {
-    Common.getCurrentTab(currentTab => {
-      var preventCloseTabUrl = Common.getPreventCloseTabUrl()
-      if (currentTab[0].url == preventCloseTabUrl) return
-      var currentDomain = (new URL(currentTab[0].url)).hostname
-      try {
-        Common.isPreventClose(value => {
-          chrome.tabs.query({
-            currentWindow: true,
-            pinned: false
-          }, tabs => {
-            [].let(it => {
-              tabs.forEach(tab => {
-                var domain = (new URL(tab.url)).hostname
-                if (!(domain != currentDomain &&
-                    ((value && tab.url != preventCloseTabUrl) ||
-                      !value))) {
-                  it.push(tab.id)
-                }
-              })
-              it.length > 0 && chrome.tabs.remove(it)
-            })
-          })
-        })
-      } catch (error) {
-        console.error(error)
-      }
-    })
-  }
-  static independent() {
-    Common.getCurrentTab(tabs => {
-      var current = tabs[0]
-      chrome.tabs.update(current.id, {
-        'pinned': !current.pinned
-      });
-    });
-  }
-}
-
-chrome.commands.onCommand.addListener(command => {
-  console.log('onCommand', command)
-  switch (command) {
-    case "toggle-pin": {
-      Commands.togglePin()
-      break
-    }
-
-    case "toggle-mute": {
-      Commands.toggleMute()
-      break
-    }
-
-    case "toShutUp": {
-      Commands.toShutUp()
-
-      break
-    }
-
-    case "duplicate": {
-      Commands.duplicate()
-
-      break
-    }
-
-    case "independent": {
-      Commands.independent()
-      break
-    }
-
-    case "newTabWithUrl": {
-      Commands.newTabWithUrl()
-      break
-    }
-
-    case "newQueryWithPasted": {
-      Commands.newQueryWithPasted()
-      break
-    }
-
-    case "newQueryWithSelected": {
-      Commands.newQueryWithSelected()
-
-      break
-    }
-
-    case "openNotepad": {
-      Commands.openNotepad()
-      break
-    }
-
-    case "killSameDomain": {
-      Commands.killSameDomain()
-      break
-    }
-
-    case "killOtherSameDomain": {
-      Commands.killOtherSameDomain()
-      break
-    }
-    case "keepSameDomain": {
-      Commands.keepSameDomain()
-
-      break
-    }
-  }
-});
