+++ src/view/popup.html
@@ -47,6 +47,4 @@
         </p>
     </div>
 </body>
-<script src="popup.js"></script>
-
 </html>
