+++ src/view/prevent_close.html
@@ -64,6 +64,5 @@
         </div>
     </div>
 </body>
-<script src="prevent_close.js?v=1"></script>
 
 </html>
